requirements.txt
pytest>=8.3.3
pytest-order>=1.2.1
pytest-html>=4.1.1
selenium>=4.24.0
webdriver-manager>=4.0.2
python-dotenv>=1.0.1